+++ src/components/navbar/navbar.css
@@ -13,6 +13,7 @@ body{
     /* backdrop-filter: blur(6px) saturate(180%); */
     -webkit-backdrop-filter: blur(6px) saturate(180%);
     background-color: transparent;
+    
 }
 .nav-logo{
     border-image: linear-gradient(319deg, #663dff 0%, #aa00ff 37%, #cc4499 100%) 1;
